Small mods to logic diagram ascii art.
Tried to switch to utf8 with an html file using better diagonals
for "diamond" conditionals, but the fonts don't align properly even
with <PRE> formatted text, the diagonal characters creating weird
horizontal offsets for all characters on the line that follow.
This seemed to be cross browser, so the technique was abandoned
in favor of this for now.